Moved exception utility methods to utility class
+review REVIEW-5566
